refactor: extract ScreenshotCapturePreflightResult.swift from PermissionAndPreflightTypes.swift
Splits the `ScreenshotCapturePreflightResult` enum (with `error`
extractor) out.

Byte-identical move of lines 63-77. Both files import Foundation.

PermissionAndPreflightTypes.swift: 83 → 67 lines. Tests: 667.
