+++ Sources/BugNarrator/Services/PermissionAndPreflightTypes.swift
@@ -60,22 +60,6 @@ struct ScreenCaptureRecoveryGuidance: Equatable {
     let message: String
 }
 
-enum ScreenshotCapturePreflightResult: Equatable {
-    case success
-    case blocked(AppError)
-    case needsUserAction(AppError)
-    case failure(AppError)
-
-    var error: AppError? {
-        switch self {
-        case .success:
-            return nil
-        case .blocked(let error), .needsUserAction(let error), .failure(let error):
-            return error
-        }
-    }
-}
-
 enum ScreenshotSelectionResult: Equatable {
     case selected(CGRect)
     case cancelled
