+++ Sources/BugNarrator/Services/ScreenshotCapturePreflightResult.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+enum ScreenshotCapturePreflightResult: Equatable {
+    case success
+    case blocked(AppError)
+    case needsUserAction(AppError)
+    case failure(AppError)
+
+    var error: AppError? {
+        switch self {
+        case .success:
+            return nil
+        case .blocked(let error), .needsUserAction(let error), .failure(let error):
+            return error
+        }
+    }
+}
